class2/pyproject.toml
[tool.poetry]
name = "class2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain-core = "^0.3.27"
langchain-google-genai = "^2.0.7"
python-dotenv = "^1.0.1"
langgraph = "^0.2.60"
langgraph-cli = {extras = ["inmem"], version = "^0.1.65"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
